+++ modules/core/02-client/keeper/client.go
@@ -16,6 +16,10 @@ import (
 func (k Keeper) CreateClient(
 	ctx sdk.Context, clientState exported.ClientState, consensusState exported.ConsensusState,
 ) (string, error) {
+	if clientState.ClientType() == exported.Localhost {
+		return "", sdkerrors.Wrapf(types.ErrInvalidClientType, "cannot create client of type: %s", clientState.ClientType())
+	}
+
 	params := k.GetParams(ctx)
 	if !params.IsAllowedClient(clientState.ClientType()) {
 		return "", sdkerrors.Wrapf(
@@ -53,7 +57,7 @@ func (k Keeper) UpdateClient(ctx sdk.Context, clientID string, clientMsg exporte
 
 	clientStore := k.ClientStore(ctx, clientID)
 
-	if status := clientState.Status(ctx, clientStore, k.cdc); status != exported.Active {
+	if status := k.GetClientStatus(ctx, clientState, clientID); status != exported.Active {
 		return sdkerrors.Wrapf(types.ErrClientNotActive, "cannot update client (%s) with status %s", clientID, status)
 	}
 
@@ -114,7 +118,7 @@ func (k Keeper) UpgradeClient(ctx sdk.Context, clientID string, upgradedClient e
 
 	clientStore := k.ClientStore(ctx, clientID)
 
-	if status := clientState.Status(ctx, clientStore, k.cdc); status != exported.Active {
+	if status := k.GetClientStatus(ctx, clientState, clientID); status != exported.Active {
 		return sdkerrors.Wrapf(types.ErrClientNotActive, "cannot upgrade client (%s) with status %s", clientID, status)
 	}