+++ modules/core/03-connection/keeper/keeper.go
@@ -1,12 +1,12 @@
 package keeper
 
 import (
+	"github.com/cometbft/cometbft/libs/log"
 	"github.com/cosmos/cosmos-sdk/codec"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-	"github.com/tendermint/tendermint/libs/log"
 
 	clienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
 	"github.com/cosmos/ibc-go/v7/modules/core/03-connection/types"
@@ -66,7 +66,7 @@ func (k Keeper) GenerateConnectionIdentifier(ctx sdk.Context) string {
 func (k Keeper) GetConnection(ctx sdk.Context, connectionID string) (types.ConnectionEnd, bool) {
 	store := ctx.KVStore(k.storeKey)
 	bz := store.Get(host.ConnectionKey(connectionID))
-	if bz == nil {
+	if len(bz) == 0 {
 		return types.ConnectionEnd{}, false
 	}
 
@@ -74,6 +74,13 @@ func (k Keeper) GetConnection(ctx sdk.Context, connectionID string) (types.Conne
 	k.cdc.MustUnmarshal(bz, &connection)
 
 	return connection, true
+}
+
+// HasConnection returns a true if the connection with the given identifier
+// exists in the store.
+func (k Keeper) HasConnection(ctx sdk.Context, connectionID string) bool {
+	store := ctx.KVStore(k.storeKey)
+	return store.Has(host.ConnectionKey(connectionID))
 }
 
 // SetConnection sets a connection to the store
@@ -106,7 +113,7 @@ func (k Keeper) GetTimestampAtHeight(ctx sdk.Context, connection types.Connectio
 func (k Keeper) GetClientConnectionPaths(ctx sdk.Context, clientID string) ([]string, bool) {
 	store := ctx.KVStore(k.storeKey)
 	bz := store.Get(host.ClientConnectionsKey(clientID))
-	if bz == nil {
+	if len(bz) == 0 {
 		return nil, false
 	}
 
@@ -127,7 +134,7 @@ func (k Keeper) SetClientConnectionPaths(ctx sdk.Context, clientID string, paths
 func (k Keeper) GetNextConnectionSequence(ctx sdk.Context) uint64 {
 	store := ctx.KVStore(k.storeKey)
 	bz := store.Get([]byte(types.KeyNextConnectionSequence))
-	if bz == nil {
+	if len(bz) == 0 {
 		panic("next connection sequence is nil")
 	}
 
@@ -187,6 +194,14 @@ func (k Keeper) GetAllConnections(ctx sdk.Context) (connections []types.Identifi
 		return false
 	})
 	return connections
+}
+
+// CreateSentinelLocalhostConnection creates and sets the sentinel localhost connection end in the IBC store.
+func (k Keeper) CreateSentinelLocalhostConnection(ctx sdk.Context) {
+	counterparty := types.NewCounterparty(exported.LocalhostClientID, exported.LocalhostConnectionID, commitmenttypes.NewMerklePrefix(k.GetCommitmentPrefix().Bytes()))
+	connectionEnd := types.NewConnectionEnd(types.OPEN, exported.LocalhostClientID, counterparty, types.ExportedVersionsToProto(types.GetCompatibleVersions()), 0)
+
+	k.SetConnection(ctx, exported.LocalhostConnectionID, connectionEnd)
 }
 
 // addConnectionToClient is used to add a connection identifier to the set of