+++ modules/core/exported/client.go
@@ -22,6 +22,12 @@ const (
 	// Wasm is used to indicate that the light client is a on-chain wasm program
 	Wasm string = "08-wasm"
 
+	// Localhost is the client type for the localhost client.
+	Localhost string = "09-localhost"
+
+	// LocalhostClientID is the sentinel client ID for the localhost client.
+	LocalhostClientID string = Localhost
+
 	// Active is a status type of a client. An active client is allowed to be used.
 	Active Status = "Active"
 
@@ -33,6 +39,9 @@ const (
 
 	// Unknown indicates there was an error in determining the status of a client.
 	Unknown Status = "Unknown"
+
+	// Unauthorized indicates that the client type is not registered as an allowed client type.
+	Unauthorized Status = "Unauthorized"
 )
 
 // ClientState defines the required common functions for light clients.