+++ testing/endpoint.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/cosmos/cosmos-sdk/baseapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
@@ -190,6 +191,7 @@ func (endpoint *Endpoint) UpgradeChain() error {
 	}
 
 	// update chain
+	baseapp.SetChainID(newChainID)(endpoint.Chain.GetSimApp().GetBaseApp())
 	endpoint.Chain.ChainID = newChainID
 	endpoint.Chain.CurrentHeader.ChainID = newChainID
 	endpoint.Chain.NextBlock() // commit changes